Consolidate Windows snapshot coverage
