+++ .github/workflows/test.yml
@@ -159,29 +159,22 @@ jobs:
           done
           exit 1
 
-      - name: Test Windows snapshots and forks
+      - name: Test Windows stopped forks
         run: |
           TEST_PATH="/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:$PATH"
-          for test_name in \
-            TestWindowsStandbyRestoreIntegration \
-            TestWindowsStandbyForkIntegration \
-            TestWindowsForkIntegration; do
-            passed=false
-            for attempt in 1 2 3; do
-              if sudo env \
-                "PATH=$TEST_PATH" \
-                "CI=true" \
-                "HYPEMAN_RUN_WINDOWS_SNAPSHOT_INTEGRATION=1" \
-                "HYPEMAN_WINDOWS_OVMF_CODE=$HYPEMAN_WINDOWS_OVMF_CODE" \
-                "HYPEMAN_WINDOWS_OVMF_VARS=$HYPEMAN_WINDOWS_OVMF_VARS" \
-                go test -count=1 -run "^${test_name}$" -timeout 2m ./lib/instances; then
-                passed=true
-                break
-              fi
-              test "$attempt" = 3 || sleep 5
-            done
-            test "$passed" = true || exit 1
+          for attempt in 1 2 3; do
+            if sudo env \
+              "PATH=$TEST_PATH" \
+              "CI=true" \
+              "HYPEMAN_RUN_WINDOWS_LIFECYCLE_INTEGRATION=1" \
+              "HYPEMAN_WINDOWS_OVMF_CODE=$HYPEMAN_WINDOWS_OVMF_CODE" \
+              "HYPEMAN_WINDOWS_OVMF_VARS=$HYPEMAN_WINDOWS_OVMF_VARS" \
+              go test -count=1 -run '^TestWindowsStoppedForkIntegration$' -timeout 2m ./lib/instances; then
+              exit 0
+            fi
+            test "$attempt" = 3 || sleep 5
           done
+          exit 1
 
       # Slash-command runs are maintainer-approved and need authenticated pulls
       # for images that are not covered by the prewarm cache.
