+++ lib/instances/windows_lifecycle_integration_linux_test.go
@@ -24,8 +24,79 @@ import (
 )
 
 func TestWindowsLifecycleIntegration(t *testing.T) {
+	manager, p, image := setupWindowsLifecycleIntegration(t)
+	ctx := context.Background()
+	source := createWindowsLifecycleInstance(t, ctx, manager, image, "windows-lifecycle-source", true)
+
+	assertWindowsNetworkReady(t, ctx, manager, source.Id, source.IP)
+	assertWindowsGuestControl(t, ctx, manager, source.Id)
+	sourceMachineID := windowsMachineID(t, ctx, manager, source.Id)
+	sourceTPMEK := windowsTPMEK(t, ctx, manager, source.Id)
+	windowsPowerShell(t, ctx, manager, source.Id, `New-Item -ItemType Directory -Force C:\ProgramData\Hypeman | Out-Null; Set-Content C:\ProgramData\Hypeman\standby.txt inherited`)
+
+	standby, err := manager.StandbyInstance(ctx, source.Id, StandbyInstanceRequest{})
+	require.NoError(t, err)
+	require.Equal(t, StateStandby, standby.State)
+	forked, err := manager.ForkInstance(ctx, source.Id, ForkInstanceRequest{
+		Name:        "windows-standby-child",
+		TargetState: StateRunning,
+	})
+	require.NoError(t, err)
+	t.Cleanup(func() { _ = deleteTestInstanceNow(context.Background(), manager, forked.Id) })
+	assert.Equal(t, source.VsockCID, forked.VsockCID, "memory-restored Windows forks retain the captured VioSock CID until cold boot")
+	assert.NotEqual(t, sourceMachineID, windowsMachineID(t, ctx, manager, forked.Id))
+	assert.Equal(t, sourceTPMEK, windowsTPMEK(t, ctx, manager, forked.Id), "memory forks inherit TPM state from the QEMU migration stream")
+	assert.Equal(t, "inherited", windowsPowerShell(t, ctx, manager, forked.Id, `Get-Content C:\ProgramData\Hypeman\standby.txt`))
+	assertIndependentFile(t, p.InstanceOVMFVars(source.Id), p.InstanceOVMFVars(forked.Id))
+	assertIndependentFile(t, p.InstanceWindowsDisk(source.Id), p.InstanceWindowsDisk(forked.Id))
+
+	stoppedFork, err := manager.StopInstance(ctx, forked.Id)
+	require.NoError(t, err)
+	require.Equal(t, StateStopped, stoppedFork.State)
+	restored, err := manager.RestoreInstance(ctx, source.Id)
+	require.NoError(t, err)
+	waitForWindowsRunning(t, ctx, manager, restored.Id)
+	assert.Equal(t, sourceMachineID, windowsMachineID(t, ctx, manager, restored.Id), "same-instance restore must preserve Windows identity")
+	assertWindowsNetworkReady(t, ctx, manager, restored.Id, source.IP)
+}
+
+func TestWindowsStoppedForkIntegration(t *testing.T) {
+	manager, p, image := setupWindowsLifecycleIntegration(t)
+	ctx := context.Background()
+	source := createWindowsLifecycleInstance(t, ctx, manager, image, "windows-stopped-fork-source", false)
+	sourceMachineID := windowsMachineID(t, ctx, manager, source.Id)
+	sourceTPMEK := windowsTPMEK(t, ctx, manager, source.Id)
+	windowsPowerShell(t, ctx, manager, source.Id, `New-Item -ItemType Directory -Force C:\ProgramData\Hypeman | Out-Null; Set-Content C:\ProgramData\Hypeman\identity.txt source`)
+
+	stopped, err := manager.StopInstance(ctx, source.Id)
+	require.NoError(t, err)
+	require.Equal(t, StateStopped, stopped.State)
+	forked, err := manager.ForkInstance(ctx, source.Id, ForkInstanceRequest{
+		Name:        "windows-stopped-child",
+		TargetState: StateRunning,
+	})
+	require.NoError(t, err)
+	t.Cleanup(func() { _ = deleteTestInstanceNow(context.Background(), manager, forked.Id) })
+	assert.NotEqual(t, sourceMachineID, windowsMachineID(t, ctx, manager, forked.Id), "fork must receive a new Windows machine identity")
+	assert.NotEqual(t, sourceTPMEK, windowsTPMEK(t, ctx, manager, forked.Id), "cold forks must initialize a fresh TPM")
+	assert.NotEqual(t, source.VsockCID, forked.VsockCID, "forks need unique vsock CIDs")
+	assertIndependentFile(t, p.InstanceOVMFVars(source.Id), p.InstanceOVMFVars(forked.Id))
+	assertIndependentFile(t, p.InstanceWindowsDisk(source.Id), p.InstanceWindowsDisk(forked.Id))
+	require.DirExists(t, p.InstanceTPMDir(forked.Id))
+
+	sourceDiskBefore, err := os.Stat(p.InstanceWindowsDisk(source.Id))
+	require.NoError(t, err)
+	windowsPowerShell(t, ctx, manager, forked.Id, `Set-Content C:\ProgramData\Hypeman\identity.txt child`)
+	assert.Equal(t, "child", windowsPowerShell(t, ctx, manager, forked.Id, `Get-Content C:\ProgramData\Hypeman\identity.txt`))
+	sourceDiskAfter, err := os.Stat(p.InstanceWindowsDisk(source.Id))
+	require.NoError(t, err)
+	assert.Equal(t, sourceDiskBefore.ModTime(), sourceDiskAfter.ModTime(), "child writes must not modify the stopped source disk")
+}
+
+func setupWindowsLifecycleIntegration(t *testing.T) (*manager, *paths.Paths, *images.Image) {
+	t.Helper()
 	if os.Getenv("HYPEMAN_RUN_WINDOWS_LIFECYCLE_INTEGRATION") != "1" {
-		t.Skip("run by the dedicated Windows networking CI gate")
+		t.Skip("run by the dedicated Windows lifecycle CI gates")
 	}
 	fixture := os.Getenv("HYPEMAN_WINDOWS_TEST_AGENT_IMAGE")
 	if fixture == "" {
@@ -33,17 +104,17 @@ func TestWindowsLifecycleIntegration(t *testing.T) {
 	}
 	if _, err := os.Stat(fixture); err != nil {
 		if os.Getenv("CI") == "true" {
-			t.Fatalf("required Windows networking fixture is missing: %s", fixture)
+			t.Fatalf("required Windows lifecycle fixture is missing: %s", fixture)
 		}
-		t.Skipf("Windows networking fixture is unavailable: %s", fixture)
+		t.Skipf("Windows lifecycle fixture is unavailable: %s", fixture)
 	}
 	acquireHeavyIO(t)
 
 	manager, dataDir := setupTestManagerForQEMU(t)
 	p := paths.New(dataDir)
-	const digestHex = "abababababababababababababababababababababababababababababababab"
+	const digestHex = "acacacacacacacacacacacacacacacacacacacacacacacacacacacacacacacac"
 	image := &images.Image{
-		Name:     "registry.example/windows/image:networking-integration",
+		Name:     "registry.example/windows/image:lifecycle-integration",
 		Digest:   "sha256:" + digestHex,
 		Platform: "windows/amd64",
 		Status:   images.StatusReady,
@@ -52,6 +123,7 @@ func TestWindowsLifecycleIntegration(t *testing.T) {
 			Base:        "registry.example/windows/base@sha256:cdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcdcd",
 			TPM:         "2.0",
 			SecureBoot:  "required",
+			BitLocker:   "disabled",
 			VirtualSize: 80 << 30,
 		},
 	}
@@ -60,28 +132,32 @@ func TestWindowsLifecycleIntegration(t *testing.T) {
 	require.NoError(t, err)
 	require.NoError(t, forkvm.CopyRegularFile(fixture, imagePath))
 	require.NoError(t, os.Chmod(imagePath, 0444))
+	return manager, p, image
+}
 
-	ctx := context.Background()
+func createWindowsLifecycleInstance(t *testing.T, ctx context.Context, manager *manager, image *images.Image, name string, networkEnabled bool) *Instance {
+	t.Helper()
 	instance, err := manager.CreateInstance(ctx, CreateInstanceRequest{
-		Name:           "windows-networking-integration",
+		Name:           name,
 		Image:          image.Name,
 		Platform:       "windows/amd64",
-		Size:           8 << 30,
+		Size:           4 << 30,
 		Vcpus:          4,
-		NetworkEnabled: true,
+		NetworkEnabled: networkEnabled,
 		Hypervisor:     hypervisor.TypeQEMU,
 	})
 	require.NoError(t, err)
 	t.Cleanup(func() { _ = deleteTestInstanceNow(context.Background(), manager, instance.Id) })
-	require.NotEmpty(t, instance.IP)
-	require.NotEmpty(t, instance.MAC)
-	require.Eventually(t, func() bool {
-		current, err := manager.GetInstance(ctx, instance.Id)
-		return err == nil && current.State == StateRunning
-	}, 4*time.Minute, time.Second)
+	waitForWindowsRunning(t, ctx, manager, instance.Id)
+	return instance
+}
 
-	assertWindowsNetworkReady(t, ctx, manager, instance.Id, instance.IP)
-	assertWindowsGuestControl(t, ctx, manager, instance.Id)
+func waitForWindowsRunning(t *testing.T, ctx context.Context, manager *manager, instanceID string) {
+	t.Helper()
+	require.Eventually(t, func() bool {
+		current, err := manager.GetInstance(ctx, instanceID)
+		return err == nil && current.State == StateRunning
+	}, 75*time.Second, 500*time.Millisecond)
 }
 
 func assertWindowsGuestControl(t *testing.T, ctx context.Context, manager *manager, instanceID string) {
@@ -100,7 +176,7 @@ func assertWindowsGuestControl(t *testing.T, ctx context.Context, manager *manag
 	require.NoError(t, err, stderr.String())
 	assert.Less(t, time.Since(jobStart), 10*time.Second)
 
-	time.Sleep(5 * time.Second)
+	time.Sleep(time.Second)
 	stdout.Reset()
 	stderr.Reset()
 	exit, err = guest.ExecIntoInstance(ctx, dialer, guest.ExecOptions{
@@ -206,4 +282,42 @@ func assertWindowsNetworkReady(t *testing.T, ctx context.Context, manager *manag
 
 	ping := exec.Command("ping", "-c", "3", "-W", "2", expectedIP)
 	require.NoError(t, ping.Run(), "allocated Windows IP did not answer ICMP")
+}
+
+func windowsMachineID(t *testing.T, ctx context.Context, manager *manager, instanceID string) string {
+	t.Helper()
+	return windowsPowerShell(t, ctx, manager, instanceID, `(Get-ItemProperty 'HKLM:\SOFTWARE\Microsoft\Cryptography').MachineGuid`)
+}
+
+func windowsTPMEK(t *testing.T, ctx context.Context, manager *manager, instanceID string) string {
+	t.Helper()
+	hash := windowsPowerShell(t, ctx, manager, instanceID, `(Get-TpmEndorsementKeyInfo -HashAlgorithm Sha256).PublicKeyHash`)
+	require.NotEmpty(t, hash, "TPM endorsement key hash")
+	return hash
+}
+
+func windowsPowerShell(t *testing.T, ctx context.Context, manager *manager, instanceID, command string) string {
+	t.Helper()
+	dialer, err := manager.GetVsockDialer(ctx, instanceID)
+	require.NoError(t, err)
+	var stdout, stderr bytes.Buffer
+	exit, err := guest.ExecIntoInstance(ctx, dialer, guest.ExecOptions{
+		Command:      []string{"powershell.exe", "-NoProfile", "-NonInteractive", "-Command", command},
+		Stdout:       &stdout,
+		Stderr:       &stderr,
+		WaitForAgent: 30 * time.Second,
+		Timeout:      30,
+	})
+	require.NoError(t, err, stderr.String())
+	require.Equal(t, 0, exit.Code, stderr.String())
+	return string(bytes.TrimSpace(stdout.Bytes()))
+}
+
+func assertIndependentFile(t *testing.T, source, fork string) {
+	t.Helper()
+	sourceInfo, err := os.Stat(source)
+	require.NoError(t, err)
+	forkInfo, err := os.Stat(fork)
+	require.NoError(t, err)
+	assert.False(t, os.SameFile(sourceInfo, forkInfo), "%s and %s must not share an inode", source, fork)
 }
